Fixed Rails 4.2 backtrace_cleaner root filter with sql-logging gem
== Primary bug fix ==

I fixed a bug that I was noticing when using this in conjunction with
other gems that leveraged `Rails.backtrace_cleaner` (ie, `sql-logging`
gem).  One of the default backtrace cleaner filters is to strip out
the application root directory from each line.  In Rails 3, it was based
on a newly allocated `Rails.root` string, but in Rails 4 it was changed
to be a frozen string generated at boot and re-used. The problem is that
the backtrace_cleaner was getting instantiated by other gems during app
initialization (and hence `Rails.root` would be `nil` affecting how the
expected trace would be filtered).

This fix sets the `@root` instance_variable of the
`Rails.backtrace_cleaner`, so while it is not optimal, it is the only way
to ensure playing well with other gems.

== Secondary bug fixes ==

The commit also fixes an issue with the implementation with
`#clean_trace`.  When the trace level was set to `:app` a silencer was
added to ensure that only app-specific lines were outputted.  However,
`Rails::BacktraceCleaner` has this behavior by default.  It was also bad
to do it this way because the `@silencers` queue was being unnecessarily
filling & calling duplicate Procs everytime `#clean_trace` was invoked.  The
commit now removes all silencers for any trace level that is not `:app`.

NB: The removal of all silencers is a bit of a shotfun approach, so I'm thinking
that it might be best just to `shift` the `@silencers`, it's just that
that solution assumes that the one that needs to be removed is the first
(which it should be until Rails modifies the default silencers...)

Finally, this commit cleans up the logged out so that there are no empty (and
misleading) `Called from: ` output.